refactor(stack): Move the model classes into a dedicated file
While at it also extract a function for the parsing of the model.
